Feat:Error Handling and Validations

+++ services/user.service.js
@@ -1,24 +1,28 @@
 const db = require('../models')
-
 const jwt = require('jsonwebtoken')
-
+const { Op } = require('sequelize')
+const { AppError } = require('../utils/error')
 const User = db.user
-
 class UserService {
-
     async createUser(userInfo) {
         try {
-            const user = await User.create(userInfo)
+            const [user, created] = await User.findOrCreate({
+                where: { [Op.or]: [{ email: userInfo.email }, { username: userInfo.username }] },
+                defaults: userInfo,
+            });
+
+            if (!created) {
+                throw new AppError('Email or Username already exists', 400);
+            }
 
             return user
         }
         catch (error) {
-            return error
+            throw error
         }
     }
 
     async login(body) {
-
         try {
             const { email, password } = body
             let user = await User.findOne({
@@ -26,11 +30,11 @@ class UserService {
             })
 
             if (!user) {
-                return 'User not found'
+                throw new AppError('User not found', 404)
             }
 
             if (user.password !== password) {
-                return 'Invalid Password'
+                throw new AppError('Incorrect Password', 400)
             }
 
             const token = jwt.sign({ id: user.id }, 'MYAPPSECRET', { expiresIn: '1d' })
@@ -38,7 +42,7 @@ class UserService {
             return token
         }
         catch (error) {
-            return error
+            throw error
         }
     }
 
@@ -49,39 +53,75 @@ class UserService {
 
             return users
         } catch (error) {
-            return error
+            throw error
         }
     }
 
-    async getUserByMail(email) {
+    async getUserByMail(email, signedInUserMail) {
         try {
+            if (email !== signedInUserMail) {
+                throw new AppError('Unauthorized to Access this Info', 401)
+            }
             const user = await User.findOne({
                 where: { email: email }
             })
 
+            if (!user) {
+                throw new AppError('User not found', 404)
+            }
+
             return user
 
         } catch (error) {
-            return error
+            throw error
         }
     }
 
     async updateUser(data, userId) {
         try {
 
-            const user = await db.user.update(data, {
+            let user = await db.user.findAll({
+                where: { id: userId }
+            })
+
+            if (!user) {
+                throw new AppError('User not found', 404)
+            }
+
+            const dataExistsInDB = await db.user.findOne({
+                where: {
+                    [Op.or]: [
+                        { email: data.email },
+                        { username: data.username }
+                    ]
+                }
+            })
+
+            if (dataExistsInDB) {
+                throw new AppError('Email or Username already Exists', 400)
+            }
+
+            user = await db.user.update(data, {
                 where: { id: userId }
             })
 
             return 'User Updated Successfully'
         }
         catch (error) {
-            return error
+            throw error
         }
     }
 
     async deleteUser(userId) {
         try {
+
+            let user = await db.user.findOne({
+                where: { id: userId }
+            })
+
+            if (!user) {
+                throw new AppError('User not found', 404)
+            }
 
             await db.user.destroy({
                 where: { id: userId }
@@ -90,7 +130,7 @@ class UserService {
             return 'User Deleted Successfully'
         }
         catch (error) {
-            return error
+            throw error
         }
     }
 }
